Fix NewFood component to use the asynchronous version of setState
React state updates are batched, and there is a potential race condition
if the current state is read as part of the update logic.
In such cases, there is an asynchronous version of setState, which allows the
update logic to use the previous state as determined by React instead of
reading the current state directly.

This commit applies that fix to relevant cases in the code base, which turns
out to be just two functions in the NewFood component.