Add 'item' property to DescribeFood component
The DescribeFood component can be used to create a new food item
or modify an existing one. In the second case, the component
state should be initialised to match the item it is modifying.

This commit adds an optional 'item' property and initialises the
state appropriately if it is supplied.