Rename UPDATE_FOOD_ITEMS action to REPLACE_FOOD_ITEMS
'REPLACE_FOOD_ITEMS' better captures the function of the action,
which is to add or overwrite entire food items.

We will soon need an UPDATE_FOOD_ITEM that merges the new partial
food item with the existing one, and it would be better if the
existing action is renamed to avoid confusion.