Rename Utility class to NestedUtility
The Utility class contains a number of convenience functions to
duplicate with a single modification to a nested property, without
mutating the original object.

I intend to create a utility method to deeply compare two objects
for equality. This is different enough from the other utility
functions that it should go in a separate class.

Since comparing objects for deep equality is more generic than
duplicating an object with a nested difference, it has a better
claim to the Utility class name.