Simplified NewRecipe to DescribeRecipe wrapper
The DescribeRecipe component contains most of the form and
logic previously contained in the NewRecipe component.
The intention is to be able to reuse that form in multiple contexts:
when creating a new recipe and when modifying an existing recipe

With this commit, the NewRecipe component simply passes a callback
to the DescribeRecipe component to create a recipe on form submission.