Create NestedUtility.retrieveFromPath
The DescribeRecipe component uses the NestedUtility class to update
a state variable whenever the corresponding input box value changes.

To handle modifying an existing recipe, we need 2-way binding:
the value of the input box needs to change whenever the state is
updated programatically. Since we already use a path string to perform
the ui-to-state binding, it is economical to parse the same string to
set the 'value' property on the input boxes (achieving the reverse binding)

NestedUtility.retrieveFromPath implements this operation