Add submodule overcooked-api
The development version of the CMS needs to interact with a local
API server. Currently, this is achieved with a script that navigates
outside the project to the directory to the directory housing the
overcooked-api project on my local machine, and running the db and
serve scripts.

Additionally, both projects default to serving on port 3000, so this
project runs a special serve script (serve-local) in the overcooked-api
project that:
  1. moves to the API server to port 4000
  2. sets a flag that allows the cross-origin requests from port 3000.

This unecessarily couples the two project configurations to each other
and to my personal development environment. Moreover, it will get
more complicated to handle the CORS configurations when adding user
authentication, which involves dealing with 3rd-party servers.

Instead, this commit adds the overcooked-api project as a submodule,
so all relevant scripts exist within the project directory.

It is also a prepatory step in proxying the CMS server to ensure
all requests can be made without reference to specific domains and ports.