Use relative urls to interact with the API server
The API server runs on port 3000 by default (4000 when the serve-local
script is used). The server is referenced with a Fully Qualified
Domain Name and port.

This necessitates CORS exceptions for the CMS server port and produces
cross-origin issues when dealing with 3rd-party servers (such as the
Github authentication server).

This commit:
  1. replaces all urls with relative urls
  2. adds a proxy to package.json so that all requests get proxied to
     the api server (and pre-flight checks treat the api server
     as the origin)