Update react proxy to handle html/text content
React uses the proxy defined in package.json for xhr requests in
development mode. However, if the proxy is set to a simple url,
it is not used for requests that accept html/text content.
This is an issue for the OAuth authentication workflow, which starts
by directing the user to the server endpoint /auth/github.

As I understand it, the react router cannot follow relative links
to external resources, and full links require hardcoding the proxy.
Therefore, I use the <a> tag with href "/auth/github".

Since requests made with <a> tags accept html/text content, the
proxy is not used by default.

This can be remedied by setting the proxy to an object, which will
be invoked for all requests that match the specified path.