Use serve-dev script as backend development server
The overcooked-api serve script has a security policy to always
send cookies over https. The prevents the backend from maintaining
a session with the development cms, which is required for user
authentication.

The serve-dev script is intended to be used during development to
bypass the secure cookie policy. This commit simply updates the
package.json to use the serve-dev script for the local backend server.