Update the proxy to target specific endpoints
During development, the cms server and the backend server are
running on different ports. In production, they will be on the
same server ( with the same url ).

The proxy is used in development to direct xhr requests
to the backend server. However, the final step in the authentication
procedure is to redirect the user back to the home page, which
should load the react Home component, instead of being proxied.

This commit updates the proxy to only apply to the backend routes
( /auth, /food and /recipes ), so other routes continue to be
passed to the backend server.

It also removes the "changeOrigin" parameter, so the http requests
still have the CMS server as the origin. This ensures that the
server redirect does not direct the browser back to the server.

Note that with this change, the authentication callback in the
GitHub configuration must now point to the CMS server instead of the
backend server.