Integrate Server auth with the redux store
The Server class provides mechanisms to retrieve the active user
and to log out. The User reducer has actions to save or clear the
active user in the store.

This commit adds functions, to be called by the User component,
that invoke Server authentication operations and then dispatch
the corresponding User redux action.