Rewrite DisplayFood to use the InteractiveTable
The DisplayFood component is a table of all food items,
each displayed by an interactive FoodItem component.

In the interests of standardisation, as well as decluttering the
edit-record functionality, each record should permit a single
action: navigate to a separate component that handles create/update
operations for food items.

This commit simplifies the DisplayFood component to be a direct
specialisation of the reusable InteractiveTable component. It also
removes the obsolete FoodItem component.

Note that although the component correctly updates the url whenever
the user interacts with a record, or decides to create a new one,
the components to be routed to are not yet implemented