Create FoodRecord component
The FoodRecord component provides a mechanism to create a new
food record or update an existing one. It achieves the same
function as the DescribeFood, NewFood and UpdateFood components
(which will all be removed in the commit) but it determines the
context from the route. It also extracts the units selector into
a separate subcomponent, simplifying the code.

This commit also updates the Container component to load the
FoodRecord component at the /cms/food routes