Rewrite DisplayRecipe to use the InteractiveTable
The DisplayRecipe component is a table of all recipes,
each displayed by an interactive RecipeItem component.

In the interests of standardisation, as well as decluttering the
edit-record functionality, each record should permit a single
action: navigate to a separate component that handles create/update
operations for recipes.

This commit simplifies the DisplayRecipe component to be a direct
specialisation of the InteractiveTable component, like the DisplayFood
component. It also removes the obsolete RecipeItem, NewRecipe,
UpdateRecipe and DescribeRecipe components

Note that although the component correctly updates the url whenever
the user interacts with a record, or decides to create a new one,
the components to be routed to are not yet implemented