Replace redux recipes array with object
The recipes are stored in redux as an array. However, now that
we are navigating to Recipe components with the ID, it would be
more convenient if they were stored as an object, referenced by
the recipe id.

This commit makes that change, and updates the store, reducers,
actions and DisplayRecipes component accordingly.