Create incomplete RecipeRecord component
The DisplayRecipes component directs users to different routes
in order to create a new recipe, or update an existing one.
Currently, there are no components loaded at those routes.

This commit creates an incomplete RecipeRecord component for
those routes, and updates the Container component accordingly.

The RecipeRecord component allows the user to see and edit
(if they are authenticated) the simple recipe properties.
To complete the component we need to:
- Create a component to display/edit the ingredients
- Create a component to display/edit the method
- Validate any changes made by the user
- Expand onSubmit to send the new data to the server