Complete RecipeRecord component
The `RecipeRecord` component allows users to view and edit
( if they are authenticated ) recipes. It currently does not
allow them to save changes to the server.

This commit confirms that the new recipes are valid, and then
requests that the server create the recipe or update an
existing one accordingly.