Close #1. Treat numerical form inputs as strings
When editing a numerical form input, if setState is called
with a null value, it gets cast to zero. This prevents the user
from directly deleting a number and writing a different one
(they could still highlight the existing value and overwrite it).

To avoid this error, we should treat all numerical inputs as
strings, and cast them back to numbers before sending the data
to the server.

The codebase adopts this approach inconsistently. This commit
ensures strings are used throughout the codebase.