Rethrow errors on all reducer actions
The codebase currently adheres to the following pattern:
- Every operation is encapsulated in a promise that first completes
  the operation and then updates the redux store
- If there is an error, the message is added to the redux store and
  the error is suppressed.
- The components ignore the result of the promise and instead rely
  on changes to the redux store to update their state

The FoodRecord and RecipeRecord components need to know if there was
an error when retrieving the record (so they can display "Not Found").

This commit changes the pattern so that all errors are also
passed to the consuming component.

For consistency, it does not suppress any error in any operation.
The components should handle them explicitly.