Suppress errors when the app loads
When the `Container` component is loaded it:
- downloads all the recipes (to be changed to just the first page)
- downloads all the food (to be changed to just the first page)
- retrieves the current user

There is nothing that the `Container` component should do if the
information cannot be loaded. This commit simply suppresses the
error.

Note that the errors are added to the redux store, so if we want
to alert the user, we should create a separate error component