Remove unnecessary layer in UnitSelector state
The `UnitSelector` state has an unnecessary layer of indirection.
It has the format { unit: { selections: [...] } } instead of
{ selections: [ ... ] }

This is due to a mistaken conclusion that in order to use the
`NestedUtility` functions to update the state, the top level cannot
be an array, since the `NestedUtility` functions only accept objects.
In fact we can simply pass the enclosing `state` object and then
extract the array from it.

This commit removes the extra layer and propagates the changes to
the `FoodRecord` component.