Query missing records in the FoodRecord component
This commit contributes towards issue #3.

The `FoodRecord` assumes that the user will arrive there by
clicking on the relevant row in the `DisplayFood` component. This
ensures the food record is already in the redux store.

However, the user can also load the page directly. In this case,
if the record is not in the redux store, the `FoodRecord`
component incorrectly displays a blank record.

This commit extends the functionality to ensure:
- an empty record is displayed if there is no id (as before)
- the record is displayed if it is available (as before)
- the record is fetched from the server if it is missing
- a loading screen is displayed while the record is fetched
- an error page is displayed if the record cannot be retrieved