Add food to redux store after retrieving recipes
When retrieving an individual recipe or a page of recipes,
the associated food items are returned as well. However, they are
simply ignored, under the assumption that all the food items are
already in the redux store. This assumption is fragile because it
depends on which components the user loads in which order.

We may as well add (or overwrite) any food items that we retrieve.
This commit modifies the recipe actions to update the food items
whenever a recipe is retrieved.

Note that creating, updating or deleting recipes are not affected.

In addition, when the app loads, all of the food items are
retrieved. The function assumes that the store is originally empty,
so any duplicate values are a sign of an error. With the changes
made in this commit, that assumption no longer holds. Therefore,
the commit also modifies the "retrieve all food items" action to
overwrite duplicate items without raising errors.