Suppress errors on server actions
The `FoodRecord` and `RecipeRecord` components currently ignore
the promise return values upon record creation, update or deletion.
Similarly, the `User` component ignores errors that occur when
signing out.

This commit explicitly suppresses any error messages.

Note that the errors are being added to the redux store,
so they can still be displayed to the user with an error component.