Build CMS under the /cms route
In production, the CMS is served under the `/cms` route. However,
it is built under the assumption that it is server at the root,
and all links are prepended with _/cms_ explicitly.

It would be more robust to configure the build process to use the
`/cms` route (using `react-scripts-ts`, this is achieved by setting
the `homepage` property of `package.json`).

More importantly, when the CMS assumes it is served from the root
of the domain, the service worker intercepts all requests, even
to the `/auth` and `/api` domains. This is interrupting the
authentication process. I suspect it may also interfere with
retrieving the latest database updates.

This commit:
- uses the latest overcooked-api submodule, which was updated
  specifically in preparation for this change
- builds the CMS knowing it will be served under `/cms`
- updates the router to prepend the `PUBLIC_URL` variable to routes
- replaces '/cms' with the `PUBLIC_URL` variable in all links