Increase resolution of Ingredient quantity
The `Ingredient` component sets the `step` property of the quantity
input box to 0.1. I thought this would simply affect how much the
quantity changes when the increase and decrease buttons are pressed,
but it actually prevents the form from accepting values with
higher resolution.

This commit increases the resolution to 0.01